+++ components/registration/OwnerRegistration.js
@@ -1,0 +1,18 @@
+import {Text} from "react-native";
+import {Component} from "react";
+
+export default class OwnerRegistration extends Component {
+
+    constructor(props) {
+        super(props);
+        const {navigation} = this.props
+
+    }
+
+
+    render() {
+        return (
+            <Text>Hello</Text>
+        );
+    }
+}
